# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SkyVoyage Weather + Travel Dashboard API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface CurrentWeather {
  city: string;
  country: string;
  temperature: number;
  condition: string;
  conditionCode?: string;
  humidity: number;
  windSpeed: number;
  uvIndex: number;
  sunrise: string;
  sunset: string;
  feelsLike: number;
  visibility: number;
  airQuality: number;
  airQualityLabel?: string;
  precipitation?: number;
}

export interface HourlyForecast {
  hour: string;
  temperature: number;
  condition: string;
  conditionCode?: string;
  precipitation: number;
  windSpeed?: number;
}

export interface DailyForecast {
  date: string;
  dayName: string;
  high: number;
  low: number;
  condition: string;
  conditionCode?: string;
  precipitation: number;
  humidity?: number;
}

export interface Destination {
  id: number;
  name: string;
  country: string;
  category: string;
  rating: number;
  travelScore: number;
  budgetLevel: string;
  bestSeason: string;
  avgTemperature: number;
  imageUrl: string;
  description: string;
  currency?: string;
  language?: string;
  timezone?: string;
  isFeatured?: boolean;
  estimatedDailyBudget?: number;
  tags?: string[];
}

export interface Trip {
  id: number;
  title: string;
  destination: string;
  /** @nullable */
  destinationId?: number | null;
  startDate: string;
  endDate: string;
  budget: number;
  status: string;
  interests: string[];
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  coverImageUrl?: string | null;
  /** @nullable */
  estimatedSpend?: number | null;
  createdAt: string;
}

export interface TripInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  destination: string;
  /** @nullable */
  destinationId?: number | null;
  startDate: string;
  endDate: string;
  budget: number;
  interests: string[];
  notes?: string;
  coverImageUrl?: string;
}

export interface TripUpdate {
  title?: string;
  destination?: string;
  startDate?: string;
  endDate?: string;
  budget?: number;
  status?: string;
  interests?: string[];
  notes?: string;
  coverImageUrl?: string;
  estimatedSpend?: number;
}

export interface ItineraryItem {
  id: number;
  tripId: number;
  date: string;
  timeOfDay: string;
  type: string;
  title: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  location?: string | null;
  duration: number;
  /** @nullable */
  cost?: number | null;
  /** @nullable */
  notes?: string | null;
}

export interface ItineraryItemInput {
  date: string;
  timeOfDay: string;
  type: string;
  /** @minLength 1 */
  title: string;
  description?: string;
  location?: string;
  duration: number;
  cost?: number;
  notes?: string;
}

export interface PackingItem {
  id: number;
  tripId: number;
  name: string;
  category: string;
  isPacked: boolean;
  isCustom: boolean;
  /** @nullable */
  quantity?: number | null;
}

export interface PackingItemInput {
  /** @minLength 1 */
  name: string;
  category: string;
  quantity?: number;
  isCustom?: boolean;
}

export interface PackingItemUpdate {
  name?: string;
  isPacked?: boolean;
  category?: string;
  quantity?: number;
}

export interface Bookmark {
  id: number;
  type: string;
  referenceId: number;
  title: string;
  /** @nullable */
  subtitle?: string | null;
  imageUrl: string;
  createdAt: string;
}

export interface BookmarkInput {
  type: string;
  referenceId: number;
  /** @minLength 1 */
  title: string;
  subtitle?: string;
  imageUrl: string;
}

export interface Notification {
  id: number;
  type: string;
  title: string;
  message: string;
  isRead: boolean;
  createdAt: string;
  /** @nullable */
  actionUrl?: string | null;
}

export interface AnalyticsSummary {
  totalTrips: number;
  destinationsVisited: number;
  upcomingTrips: number;
  totalSpend: number;
  savedDestinations: number;
  mostVisitedRegion?: string;
  avgTripDuration?: number;
}

export interface TemperaturePoint {
  month: string;
  high: number;
  low: number;
  rainfall: number;
}

export interface DestinationStat {
  name: string;
  country: string;
  visits: number;
  score: number;
  imageUrl?: string;
}

export interface AiSuggestion {
  id: number;
  category: string;
  message: string;
  confidence: number;
  /** @nullable */
  actionLabel?: string | null;
  /** @nullable */
  actionUrl?: string | null;
  icon?: string;
}

export type GetCurrentWeatherParams = {
/**
 * City name to search
 */
city?: string;
};

export type GetHourlyForecastParams = {
city?: string;
};

export type GetWeeklyForecastParams = {
city?: string;
};

export type GetDestinationsParams = {
category?: string;
search?: string;
budget?: string;
};

export type GetTemperatureTrendsParams = {
city?: string;
};

export type GetAiSuggestionsParams = {
/**
 * Context for suggestions (e.g. destination or season)
 */
context?: string;
};