# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SkyVoyage Weather + Travel Dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /destinations:
    get:
      operationId: getDestinations
      tags: [destinations]
      summary: List destinations with optional filters
      parameters:
        - in: query
          name: category
          schema: { type: string }
        - in: query
          name: search
          schema: { type: string }
        - in: query
          name: budget
          schema: { type: string }
      responses:
        "200":
          description: List of destinations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Destination"
  /destinations/{id}:
    get:
      operationId: getDestination
      tags: [destinations]
      summary: Get a single destination by ID
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Destination detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Destination"
...
  /trips:
    get:
      operationId: getTrips
      tags: [trips]
      summary: List all trips
      responses:
        "200":
          description: List of trips
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Trip"
    post:
      operationId: createTrip
      tags: [trips]
      summary: Create a new trip
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TripInput"
      responses:
        "201":
          description: Created trip
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
  /trips/{id}:
    get:
      operationId: getTrip
      tags: [trips]
      summary: Get a trip by ID
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Trip detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
    patch:
      operationId: updateTrip
      tags: [trips]
      summary: Update a trip
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TripUpdate"
      responses:
        "200":
          description: Updated trip
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
    delete:
      operationId: deleteTrip
      tags: [trips]
      summary: Delete a trip
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ─── ITINERARY ──────────────────────────────────────────────────────────────
  /trips/{tripId}/itinerary:
    get:
      operationId: getItinerary
      tags: [itinerary]
      summary: Get itinerary items for a trip
      parameters:
        - in: path
          name: tripId
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Itinerary items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ItineraryItem"
    post:
      operationId: createItineraryItem
      tags: [itinerary]
      summary: Add an item to the itinerary
      parameters:
        - in: path
          name: tripId
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ItineraryItemInput"
      responses:
        "201":
          description: Created itinerary item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItineraryItem"
  /trips/{tripId}/itinerary/{itemId}:
    delete:
      operationId: deleteItineraryItem
      tags: [itinerary]
      summary: Remove an itinerary item
      parameters:
        - in: path
          name: tripId
          required: true
          schema: { type: integer }
        - in: path
          name: itemId
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ─── PACKING ────────────────────────────────────────────────────────────────
  /trips/{tripId}/packing:
    get:
      operationId: getPackingList
      tags: [packing]
      summary: Get packing list for a trip
      parameters:
        - in: path
          name: tripId
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Packing list items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PackingItem"
    post:
      operationId: createPackingItem
      tags: [packing]
      summary: Add a packing item
      parameters:
        - in: path
          name: tripId
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PackingItemInput"
      responses:
        "201":
          description: Created packing item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PackingItem"
  /trips/{tripId}/packing/{itemId}:
    patch:
      operationId: updatePackingItem
      tags: [packing]
      summary: Toggle packed status or update item
      parameters:
        - in: path
          name: tripId
          required: true
          schema: { type: integer }
        - in: path
          name: itemId
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PackingItemUpdate"
      responses:
        "200":
          description: Updated item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PackingItem"
    delete:
      operationId: deletePackingItem
      tags: [packing]
      summary: Delete a packing item
      parameters:
        - in: path
          name: tripId
          required: true
          schema: { type: integer }
        - in: path
          name: itemId
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ─── BOOKMARKS ──────────────────────────────────────────────────────────────
  /bookmarks:
    get:
      operationId: getBookmarks
      tags: [bookmarks]
      summary: Get all bookmarks
      responses:
        "200":
          description: List of bookmarks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Bookmark"
    post:
      operationId: createBookmark
      tags: [bookmarks]
      summary: Bookmark a destination or trip
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookmarkInput"
      responses:
        "201":
          description: Created bookmark
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bookmark"
  /bookmarks/{id}:
    delete:
      operationId: deleteBookmark
      tags: [bookmarks]
      summary: Remove a bookmark
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ─── NOTIFICATIONS ──────────────────────────────────────────────────────────
...
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a notification as read
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Updated notification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"

  # ─── ANALYTICS ──────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Destination:
      type: object
      required: [id, name, country, category, rating, travelScore, budgetLevel, bestSeason, avgTemperature, imageUrl, description]
      properties:
        id: { type: integer }
        name: { type: string }
        country: { type: string }
        category: { type: string }
        rating: { type: number }
        travelScore: { type: integer }
        budgetLevel: { type: string }
        bestSeason: { type: string }
        avgTemperature: { type: number }
        imageUrl: { type: string }
        description: { type: string }
        currency: { type: string }
        language: { type: string }
        timezone: { type: string }
        isFeatured: { type: boolean }
        estimatedDailyBudget: { type: number }
        tags:
          type: array
          items: { type: string }

    # ─── TRIPS ────────────────────────────────────────────────────────────────
    Trip:
      type: object
      required: [id, title, destination, startDate, endDate, budget, status, interests, createdAt]
      properties:
        id: { type: integer }
        title: { type: string }
        destination: { type: string }
        destinationId: { type: ["integer", "null"] }
        startDate: { type: string }
        endDate: { type: string }
        budget: { type: number }
        status: { type: string }
        interests:
          type: array
          items: { type: string }
        notes: { type: ["string", "null"] }
        coverImageUrl: { type: ["string", "null"] }
        estimatedSpend: { type: ["number", "null"] }
        createdAt: { type: string }
    TripInput:
      type: object
      required: [title, destination, startDate, endDate, budget, interests]
      properties:
        title: { type: string, minLength: 1 }
        destination: { type: string, minLength: 1 }
        destinationId: { type: ["integer", "null"] }
        startDate: { type: string }
        endDate: { type: string }
        budget: { type: number }
        interests:
          type: array
          items: { type: string }
        notes: { type: string }
        coverImageUrl: { type: string }
    TripUpdate:
      type: object
      properties:
        title: { type: string }
        destination: { type: string }
        startDate: { type: string }
        endDate: { type: string }
        budget: { type: number }
        status: { type: string }
        interests:
          type: array
          items: { type: string }
        notes: { type: string }
        coverImageUrl: { type: string }
        estimatedSpend: { type: number }

    # ─── ITINERARY ────────────────────────────────────────────────────────────
    ItineraryItem:
      type: object
      required: [id, tripId, date, timeOfDay, type, title, duration]
      properties:
        id: { type: integer }
        tripId: { type: integer }
        date: { type: string }
        timeOfDay: { type: string }
        type: { type: string }
        title: { type: string }
        description: { type: ["string", "null"] }
        location: { type: ["string", "null"] }
        duration: { type: integer }
        cost: { type: ["number", "null"] }
        notes: { type: ["string", "null"] }
    ItineraryItemInput:
      type: object
      required: [date, timeOfDay, type, title, duration]
      properties:
        date: { type: string }
        timeOfDay: { type: string }
        type: { type: string }
        title: { type: string, minLength: 1 }
        description: { type: string }
        location: { type: string }
        duration: { type: integer }
        cost: { type: number }
        notes: { type: string }

    # ─── PACKING ──────────────────────────────────────────────────────────────
    PackingItem:
      type: object
      required: [id, tripId, name, category, isPacked, isCustom]
      properties:
        id: { type: integer }
        tripId: { type: integer }
        name: { type: string }
        category: { type: string }
        isPacked: { type: boolean }
        isCustom: { type: boolean }
        quantity: { type: ["integer", "null"] }
    PackingItemInput:
      type: object
      required: [name, category]
      properties:
        name: { type: string, minLength: 1 }
        category: { type: string }
        quantity: { type: integer }
        isCustom: { type: boolean }
    PackingItemUpdate:
      type: object
      properties:
        name: { type: string }
        isPacked: { type: boolean }
        category: { type: string }
        quantity: { type: integer }

    # ─── BOOKMARKS ────────────────────────────────────────────────────────────
    Bookmark:
      type: object
      required: [id, type, referenceId, title, imageUrl, createdAt]
      properties:
        id: { type: integer }
        type: { type: string }
        referenceId: { type: integer }
        title: { type: string }
        subtitle: { type: ["string", "null"] }
        imageUrl: { type: string }
        createdAt: { type: string }
    BookmarkInput:
      type: object
      required: [type, referenceId, title, imageUrl]
      properties:
        type: { type: string }
        referenceId: { type: integer }
        title: { type: string, minLength: 1 }
        subtitle: { type: string }
        imageUrl: { type: string }

    # ─── NOTIFICATIONS ────────────────────────────────────────────────────────
    Notification:
      type: object
      required: [id, type, title, message, isRead, createdAt]
      properties:
        id: { type: integer }
        type: { type: string }
        title: { type: string }
        message: { type: string }
        isRead: { type: boolean }
        createdAt: { type: string }
        actionUrl: { type: ["string", "null"] }

    # ─── ANALYTICS ────────────────────────────────────────────────────────────
...