# lib/api-zod/src/generated/types/bookmark.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SkyVoyage Weather + Travel Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface Bookmark {
  id: number;
  type: string;
  referenceId: number;
  title: string;
  /** @nullable */
  subtitle?: string | null;
  imageUrl: string;
  createdAt: string;
}