# lib/api-zod/src/generated/types/destination.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SkyVoyage Weather + Travel Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface Destination {
  id: number;
  name: string;
  country: string;
  category: string;
  rating: number;
  travelScore: number;
  budgetLevel: string;
  bestSeason: string;
  avgTemperature: number;
  imageUrl: string;
  description: string;
  currency?: string;
  language?: string;
  timezone?: string;
  isFeatured?: boolean;
  estimatedDailyBudget?: number;
  tags?: string[];
}