# lib/api-zod/src/generated/types/itineraryItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SkyVoyage Weather + Travel Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface ItineraryItem {
  id: number;
  tripId: number;
  date: string;
  timeOfDay: string;
  type: string;
  title: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  location?: string | null;
  duration: number;
  /** @nullable */
  cost?: number | null;
  /** @nullable */
  notes?: string | null;
}